Split names into file per class
