+++ lib/mocha/impersonating_any_instance_name.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Mocha
+  class ImpersonatingAnyInstanceName
+    def initialize(klass)
+      @klass = klass
+    end
+
+    def mocha_inspect
+      "#<AnyInstance:#{@klass.mocha_inspect}>"
+    end
+  end
+end
