+++ lib/mocha/mock.rb
@@ -4,7 +4,7 @@ require 'ruby2_keywords'
 require 'mocha/expectation'
 require 'mocha/expectation_list'
 require 'mocha/invocation'
-require 'mocha/names'
+require 'mocha/default_name'
 require 'mocha/default_receiver'
 require 'mocha/method_matcher'
 require 'mocha/parameters_matcher'
